# api/openapi.yaml
openapi: 3.1.0
info:
  title: memini
  description: A Kubernetes-ready memory service for AI agents.
  version: 0.1.0
  license:
    name: AGPL-3.0-only
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /v1/handshake:
    post:
      operationId: handshake
      tags: [memory]
      summary: Resolve namespace, identity, and behavioral settings from client-supplied project facts
      description: >
        The client→server handshake: the client sends what it knows about the
        project (git remote/toplevel/cwd, an optional agent suffix, and any
        namespace it already has an opinion about) plus its own name/version;
        the server resolves the effective namespace (an explicit pin beats
        MEMINI_NAMESPACE beats a declared value beats derivation from the
        project facts beats the caller's key-level default beats the server
        default), the caller's identity, the fully-merged ClientSettings
        (built-in defaults, overridden by the server's global defaults,
        overridden by any per-key settings), and the read-set the resolved
        namespace draws from. This is the single source of truth for
        namespace resolution — clients no longer derive or cache it locally.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HandshakeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HandshakeResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/self:
    get:
      operationId: getSelf
      tags: [memory]
      summary: The caller's identity and fully-merged behavioral settings
      description: >
        A lighter-weight refresh than /v1/handshake for a client that already
        has a resolved namespace and only wants current identity/settings —
        no project facts to resend, and no read_set/pin/server info.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SelfResponse" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/self/settings:
    put:
      operationId: putSelfSettings
      tags: [memory]
      summary: Replace the caller's own per-key behavioral settings
      description: >
        Full-replace semantics: the body REPLACES the settings stored on
        the API key the request authenticated with. Only the fields
        explicitly set in the body are stored; a field omitted from the
        body returns to inheriting (the server's global defaults, then
        the built-ins). Clients are expected to GET /v1/self, edit, and
        PUT the result back — there is no merge/patch variant. Requires a
        NAMED key: the admin env key and dev mode authenticate with no
        principal, so there is no "self" to update — 403, use
        PUT /v1/settings/defaults for the global layer instead. 409 for a
        MEMINI_API_KEYS_FILE-sourced key (its settings come from that file,
        immutable at runtime — same convention as PATCH /v1/keys/{name}).
        501 against a storage backend that cannot persist per-key settings.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientSettings" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SelfResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
...
  /v1/pins:
    get:
      operationId: listPins
      tags: [memory]
      summary: List explicit project→namespace pins
      description: >
        Open to every credential class: the project map is machine-wide
        derivation state, not scoped to one namespace, and any caller's
        handshake already reveals whichever pin matches its project facts.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectMapListResponse" }
        "501": { $ref: "#/components/responses/Error" }
    put:
      operationId: putPin
      tags: [memory]
      summary: Create or replace an explicit project→namespace pin
      description: >
        Any caller may pin — admin key, dev mode, or a named table/file key.
        Writes are activity-logged with their author (event kind "pin");
        that audit trail is the control, not an authorization gate. Pins
        are keyed by remote_url (canonicalized) and/or toplevel_path — at
        least one must be given, 400 otherwise. A pin beats every other
        namespace_source, including MEMINI_NAMESPACE, at handshake time.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectMapPutRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectMapEntry" }
        "400": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deletePin
      tags: [memory]
      summary: Delete an explicit project→namespace pin
      description: >
        Any caller may unpin — same authorization as putPin, with
        the delete activity-logged the same way (event kind "unpin").
        remote_url and/or toplevel_path identify the pin to remove — at
        least one must be given, 400 otherwise.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectMapDeleteRequest" }
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
    Tier:
      type: string
      enum: [working, episodic, semantic, procedural]
...
    ReadSetOrigin:
      type: string
      description: >
        Why a namespace is in the read-set: "primary" is the request
        namespace (and its subtree, when expanded), "ancestor" is a
        path-prefix cascade leg, "home" is the caller's personal namespace,
        "link" is a stored namespace link, and "call" is an explicit per-call
        namespace.
      enum: [primary, ancestor, home, link, call]
    ReadSetEntryItem:
      type: object
      required: [namespace, origin]
      properties:
        namespace: { type: string }
        origin: { $ref: "#/components/schemas/ReadSetOrigin" }
        tiers:
          type: array
          items: { $ref: "#/components/schemas/Tier" }
          description: >
            Tier restriction applied to this namespace; omitted means the
            request's own tier filter, unrestricted beyond that.
...
    CallerIdentity:
      type: object
      required: [authenticated]
      description: Who the request authenticated as, independent of any resolved namespace.
      properties:
        authenticated: { type: boolean }
        key_name:
          type: string
          description: >
            Name of the API key that authenticated the request; absent for
            the admin key or dev mode (no named principal — see
            requireAdminOrDev's doc for the same distinction on /v1/keys).
        home:
          type: string
          description: The key's bound home namespace, if any.
        default_namespace:
          type: string
          description: The key's bound default namespace, if any.
    ClientSettings:
      type: object
      description: >
        Behavioral/injection settings, resolved by merging built-in defaults
        with any server global defaults and any per-key override. Every
        field is optional in the schema — absent means "inherit from the
        next layer down" — but a fully resolved ClientSettings (as returned
        by /v1/handshake, /v1/self, and GET /v1/settings/defaults) always
        carries every field.
      properties:
        capture_turns:
          type: boolean
          default: true
          description: Capture each user→assistant turn as episodic memory.
        session_digest:
          type: boolean
          default: true
          description: Record a session-end/stop/pre-compact digest memory.
        inline_extract:
          type: boolean
          default: true
          description: >
            Inject the directive asking the agent to save durable facts via
            memory_remember.
        auto_save:
          type: boolean
          default: true
          description: Periodically nudge the agent to persist durable memories.
        auto_save_interval:
          type: integer
          minimum: 1
          default: 10
          description: User-message interval between auto-save nudges.
        inject_briefing_pinned:
          type: integer
          minimum: 0
          default: 5
          description: Max pinned memories in the session-start briefing.
        inject_briefing_facts:
          type: integer
          minimum: 0
          default: 5
          description: Max durable semantic facts in the session-start briefing.
        inject_briefing_procedures:
          type: integer
          minimum: 0
          default: 5
          description: Max procedural how-tos in the session-start briefing.
        inject_briefing_recent:
          type: integer
          minimum: 0
          default: 3
          description: Max recent episodic entries in the session-start briefing.
        inject_briefing_max_tok:
          type: integer
          minimum: 0
          default: 0
          description: Hard ceiling on briefing injection tokens; 0 is uncapped.
        inject_pretool_items:
          type: integer
          minimum: 0
          default: 3
          description: Max recalled items injected per file on PreToolUse.
        inject_pretool_max_tok:
          type: integer
          minimum: 0
          default: 0
          description: Hard ceiling on per-tool injection tokens; 0 is uncapped.
        inject_pretool_min_score:
          type: number
          minimum: 0
          default: 0
          description: Floor on the fused score (>=) for a PreToolUse injection.
        inject_pretool_tools:
          type: array
          items: { type: string }
          default: [Read, Write, Edit, Glob, Grep]
          description: Tool-name allowlist that triggers a PreToolUse injection.
        inject_labels:
          type: array
          default: []
          items:
            type: string
            enum: [tier, confidence, age, reason]
            # Codegen metadata only (no wire effect): without explicit names,
            # oapi-codegen would mint a bare `Tier` constant for the "tier"
            # value, colliding with the Tier schema's generated Go type.
            x-enum-varnames:
              [InjectLabelTier, InjectLabelConfidence, InjectLabelAge, InjectLabelReason]
          description: Which annotation labels to render alongside an injected memory.
        recall:
          type: boolean
          default: true
          description: Enable recall-driven injection at all.
        capture:
          type: boolean
          default: true
          description: Enable capture (turns/digests) at all.
        recall_limit:
          type: integer
          minimum: 0
          default: 3
          description: Max memories per recall call.
        inject_recall_max_tok:
          type: integer
          minimum: 0
          default: 0
          description: Hard ceiling on recall injection tokens; 0 is uncapped.
        inject_recall_min_score:
          type: number
          minimum: 0
          default: 0
          description: Floor on the fused score (>=) for a recall injection.
        min_capture_chars:
          type: integer
          minimum: 0
          default: 0
          description: Minimum content length worth bothering to capture a turn.
        namespace_scope:
          type: string
          enum: [repo, owner_repo]
          default: repo
          description: >
            "repo" derives the namespace from the bare repo name; "owner_repo"
            disambiguates same-named repos across owners with an
            owner-repo slug (owner + "-" + repo).
        namespace_prefix:
          type: string
          default: ""
          description: Namespace path prepended ahead of the derived/declared namespace.
...
    SelfResponse:
      type: object
      required: [identity, settings, settings_sources]
      properties:
        identity: { $ref: "#/components/schemas/CallerIdentity" }
        settings:
          allOf:
            - $ref: "#/components/schemas/ClientSettings"
          description: Fully resolved — every field present.
        settings_sources:
          type: object
          description: Per-field provenance for `settings`, keyed by the same field names.
          additionalProperties:
            type: string
            enum: [default, global, key]
    HandshakeRequest:
      type: object
      required: [project]
      description: >
        Client→server handshake input: what the client currently knows about
        the project and about itself. project.cwd_basename is the only field
        every caller can always supply (even a bare directory with no git
        repo); everything else sharpens the resolution.
      properties:
        project:
          type: object
          required: [cwd_basename]
          properties:
            remote_url:
              type: string
              description: >
                Raw git remote URL, exactly as `git remote get-url origin`
                reports it — unnormalized.
            toplevel_path:
              type: string
              description: >
                Absolute git toplevel directory — the machine-local pin key
                (path:<toplevel_path>).
            toplevel_basename:
              type: string
              description: Basename of toplevel_path, for the toplevel fallback derivation.
            cwd_basename:
              type: string
              description: >
                Basename of the working directory — the last-resort
                derivation fallback, always present.
            agent:
              type: string
              description: >
                Per-agent suffix (e.g. "reviewer"), sanitized and appended as
                a nested namespace segment.
            env_namespace:
              type: string
              description: >
                The client's MEMINI_NAMESPACE, when set. Sent so a pin can
                still beat it server-side — the client cannot make that call
                itself without knowing whether a pin exists.
            declared_namespace:
              type: string
              description: >
                For gateway/integration callers with no meaningful cwd (a
                webhook relay, a CI job) to declare their namespace directly.
                Wins over derivation but not over a pin or env_namespace.
        client:
          type: object
          description: Caller identification, for logging/diagnostics only.
          properties:
            name: { type: string }
            version: { type: string }
    HandshakeResponse:
      type: object
      required:
        [namespace, namespace_source, identity, settings, settings_sources, read_set, server]
      properties:
        namespace:
          type: string
          description: The resolved namespace this caller should use for every subsequent request.
        namespace_source:
          type: string
          description: Which rule resolved `namespace`, highest precedence first.
          enum: [pin, env, declared, remote, toplevel, cwd, key_default, server_default]
        pin:
          type: object
          description: Present only when namespace_source is "pin".
          required: [key, updated_at]
          properties:
            key:
              type: string
              description: >
                The project_map key that matched
                (remote:<canonical-remote> or path:<toplevel_path>).
            note: { type: string }
            created_by:
              type: string
              description: >
                Name of the API key that created the pin, when known —
                same semantics as ProjectMapEntry.created_by (absent for
                a pin created by the admin key or in dev mode, which
                carry no named principal).
            updated_at: { type: string, format: date-time }
        identity: { $ref: "#/components/schemas/CallerIdentity" }
        settings:
          allOf:
            - $ref: "#/components/schemas/ClientSettings"
          description: Fully resolved — every field present.
        settings_sources:
          type: object
          description: Per-field provenance for `settings`, keyed by the same field names.
          additionalProperties:
            type: string
            enum: [default, global, key]
        read_set:
          type: array
          description: >
            The read-set `namespace` resolves to (same shape as GET
            /v1/namespaces/readset).
          items: { $ref: "#/components/schemas/ReadSetEntryItem" }
        server:
          type: object
          required: [version, default_namespace]
          properties:
            version: { type: string }
            default_namespace: { type: string }
    ProjectMapEntry:
      type: object
      required: [key, namespace, created_at, updated_at]
      description: >
        One explicit project→namespace pin, keyed by remote_url
        (canonicalized) or toplevel_path.
      properties:
        key:
          type: string
          description: 'The project_map key: "remote:<canonical-remote>" or "path:<abs-toplevel>".'
        namespace: { type: string }
        note: { type: string }
        created_by:
          type: string
          description: Name of the API key that created the pin, when known.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProjectMapListResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/ProjectMapEntry" }
    ProjectMapPutRequest:
      type: object
      required: [namespace]
      description: >
        At least one of remote_url/toplevel_path is required (400 if
        neither is given) — that is the key fact the pin is stored under.
      properties:
        namespace: { type: string }
        remote_url:
          type: string
          description: >
            Raw git remote URL; canonicalized server-side into the
            remote:<canonical> key.
        toplevel_path:
          type: string
          description: Absolute git toplevel path; stored as the path:<toplevel_path> key.
        note: { type: string }
    ProjectMapDeleteRequest:
      type: object
      description: >
        At least one of remote_url/toplevel_path is required (400 if
        neither is given) to identify the pin to delete.
      properties:
        remote_url: { type: string }
        toplevel_path: { type: string }